dataset/requirements.txt
nltk
bert-extractive-summarizer
sentence-transformers
torch
gpt4all
tqdm
zstandard
textsum
unidecode
# pycryptodome
pycryptodomex
webdavclient3

# llamav1 summarizer
langchain
huggingface_hub
llama-cpp-python
# See https://github.com/abetlen/llama-cpp-python/issues/1285#issuecomment-2006153790 :
# force build for metal: CMAKE_ARGS="-DLLAMA_METAL_EMBED_LIBRARY=ON -DLLAMA_METAL=on" pip install -U llama-cpp-python --no-cache-dir
